+++ dist-deno/schemaClasses/StringSchema.ts
@@ -1,8 +1,8 @@
-import * as converter from "../appliers/converter.ts";
 import * as ifEmptyString from "../appliers/ifEmptyString.ts";
 import * as ifNull from "../appliers/ifNull.ts";
 import * as ifUndefined from "../appliers/ifUndefined.ts";
 import * as only from "../appliers/only.ts";
+import * as transform from "../appliers/transform.ts";
 import * as maxLength from "../appliers/string/maxLength.ts";
 import * as minLength from "../appliers/string/minLength.ts";
 import * as pattern from "../appliers/string/pattern.ts";
@@ -12,7 +12,7 @@ import { BaseSchema } from "./BaseSchema.ts";
 export const STRING = {
     PATTERN: pattern.PATTERN
 } as const;
-export type RulesForString = converter.Rules<string> & ifEmptyString.Rules<string> & ifNull.Rules<string> & ifUndefined.Rules<string> & only.Rules<string> & type.Rules & trims.Rules & minLength.Rules & maxLength.Rules & pattern.Rules;
+export type RulesForString = transform.Rules<string> & ifEmptyString.Rules<string> & ifNull.Rules<string> & ifUndefined.Rules<string> & only.Rules<string> & type.Rules & trims.Rules & minLength.Rules & maxLength.Rules & pattern.Rules;
 export class StringSchema<Tx = never> extends BaseSchema<string | Tx> {
     constructor(rules: RulesForString) {
         super(rules, [
@@ -25,7 +25,7 @@ export class StringSchema<Tx = never> extends BaseSchema<string | Tx> {
             minLength.applyTo,
             maxLength.applyTo,
             pattern.applyTo,
-            converter.applyTo,
+            transform.applyTo,
         ]);
     }
 }