+++ dist-deno/schemas/object.ts
@@ -1,6 +1,6 @@
+import { ObjectSchema, RulesForObject } from "../schemaClasses/ObjectSchema.ts";
 import { NullableRules, UndefinableRules } from "../libs/publicTypes.ts";
 import { SchemaObject } from "../libs/types.ts";
-import { ObjectSchema, RulesForObject } from "../schemaClasses/ObjectSchema.ts";
 /** schema for object or null or undefined */
 export function object<S extends SchemaObject>(options: RulesForObject<S> & NullableRules & UndefinableRules): ObjectSchema<S, null | undefined>;
 /** schema for object or undefined */