+++ src/appliers/array/each.ts
@@ -1,5 +1,7 @@
 import {isArray, Key, Values} from "../../libs/types";
+
 import {ValueSchemaError} from "../../libs/ValueSchemaError";
+
 import {BaseSchema} from "../../schemaClasses/BaseSchema";
 
 type Each<T> = {