+++ src/schemaClasses/DateSchema.ts
@@ -1,12 +1,13 @@
-import * as converter from "../appliers/converter";
 import * as ifEmptyString from "../appliers/ifEmptyString";
 import * as ifNull from "../appliers/ifNull";
 import * as ifUndefined from "../appliers/ifUndefined";
+import * as transform from "../appliers/transform";
+
 import * as iso8601 from "../appliers/date/iso8601";
 import * as maxValue from "../appliers/date/maxValue";
 import * as minValue from "../appliers/date/minValue";
-import * as unixtime from "../appliers/date/unixtime";
 import * as type from "../appliers/date/type";
+import * as unixtime from "../appliers/date/unixtime";
 
 import {BaseSchema} from "./BaseSchema";
 
@@ -15,7 +16,7 @@ export const DATE = {
 } as const;
 
 export type RulesForDate =
-	converter.Rules<Date> &
+	transform.Rules<Date> &
 	ifUndefined.Rules<Date> &
 	ifEmptyString.Rules<Date> &
 	ifNull.Rules<Date> &
@@ -37,7 +38,7 @@ export class DateSchema<Tx = never> extends BaseSchema<Date | Tx>
 			maxValue.applyTo,
 			minValue.applyTo,
 			type.applyTo,
-			converter.applyTo,
+			transform.applyTo,
 		]);
 	}
 }