+++ src/schemaClasses/ObjectSchema.ts
@@ -1,7 +1,8 @@
-import * as converter from "../appliers/converter";
-import * as ifUndefined from "../appliers/ifUndefined";
 import * as ifEmptyString from "../appliers/ifEmptyString";
 import * as ifNull from "../appliers/ifNull";
+import * as ifUndefined from "../appliers/ifUndefined";
+import * as transform from "../appliers/transform";
+
 import * as schema from "../appliers/object/schema";
 import * as type from "../appliers/object/type";
 
@@ -10,7 +11,7 @@ import {AnyObject, ObjectTypeOf, SchemaObject} from "../libs/types";
 import {BaseSchema} from "./BaseSchema";
 
 export type RulesForObject<S> =
-	converter.Rules<AnyObject> &
+	transform.Rules<AnyObject> &
 	ifUndefined.Rules<AnyObject> &
 	ifEmptyString.Rules<AnyObject> &
 	ifNull.Rules<AnyObject> &
@@ -27,7 +28,7 @@ export class ObjectSchema<S extends SchemaObject, Tx = never> extends BaseSchema
 			ifEmptyString.applyTo,
 			schema.applyTo,
 			type.applyTo,
-			converter.applyTo,
+			transform.applyTo,
 		]);
 	}
 }