+++ test/schemas/email.test.ts
@@ -1,5 +1,5 @@
-import vs from "value-schema";
 import {describe, expect, it} from "@jest/globals";
+import vs from "value-schema";
 
 {
 	describe("isUndefined", testIfUndefined);