+++ test/schemas/numericString.test.ts
@@ -1,5 +1,5 @@
-import vs from "value-schema";
 import {describe, expect, it} from "@jest/globals";
+import vs from "value-schema";
 
 {
 	describe("type", testType);
@@ -12,7 +12,7 @@ import {describe, expect, it} from "@jest/globals";
 	describe("checksum (Luhn)", testChecksumLuhn);
 	describe("checksum (Modulus 10 / Weight 3:1)", testChecksumModulus10Weight31);
 	describe("checksum (Others)", testChecksumOthers);
-	describe("converter", testConverter);
+	describe("transform", testTransform);
 }
 
 /**
@@ -310,15 +310,15 @@ function testChecksumOthers(): void
 }
 
 /**
- * converter
+ * transform
  */
-function testConverter(): void
+function testTransform(): void
 {
 	it("should be adjusted", () =>
 	{
 		expect(
 			vs.numericString({
-				converter: (value, fail) =>
+				transform: (value, fail) =>
 				{
 					if(value.length >= 9)
 					{
@@ -334,7 +334,7 @@ function testConverter(): void
 		expect(() =>
 		{
 			vs.string({
-				converter: (value, fail) =>
+				transform: (value, fail) =>
 				{
 					if(value.length >= 9)
 					{
@@ -343,6 +343,6 @@ function testConverter(): void
 					return value.padStart(8, "0");
 				},
 			}).applyTo("123456789");
-		}).toThrow(vs.RULE.CONVERTER);
+		}).toThrow(vs.RULE.TRANSFORM);
 	});
 }